python/backend/requirements.txt
anyio==3.7.0
awxkit==22.3.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.0
dnspython==2.3.0
fastapi==0.95.2
Flask==2.3.2
Flask-SocketIO==5.3.4
h11==0.14.0
httptools==0.5.0
hvac==1.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.22.0
paramiko==3.2.0
pycparser==2.21
pydantic==1.10.8
pyhcl==0.4.4
PyJWT==2.7.0
pymongo==4.3.3
PyNaCl==1.5.0
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
sse-starlette==1.6.1
starlette==0.27.0
typing_extensions==4.6.2
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
